Add files via upload
This project takes in encoded messages and outputs the decoded version. The code uses the known frequencies of letters and matches them to the frequencies of the letters from the encoded text. (e t a o i n s r h l d c u m f p g w y b v k x j q z) this is the order of frequency of the letters in the English language. Based on this, if an encoded text has 100 r, 82 v, 78 o, and so on, r would be mapped to e, v - t, o - a and so on. There are some problems that arise, and decoded messages generally are not perfect. You can, however, generally, make out meaning from the decoded messages.